reflection/requirements.txt
numpy
pyrealsense2
torch
torchvision
opencv-python
pycocotools
python-socketio
requests
eventlet
dataclasses
av
scipy
setuptools==41.0.0